Update 05.07 중력을 적용한 점프 및 떨어지기 재구현 기존 점프애니메이션 삭제(인폼이 아님) 중력가속도를 적용하여 떨어지는 속도가 자연스러움 몬스터 추가 생성 몬스터 시야가 오브젝트와 충돌시 방향을 반대로 바꿔버림 -> 몬스터가 오브젝트를 뚫고 지나가지 않게 파티클이펙트 진행중 현재 몬스터 타격시 이펙트 터지게는 되지만 2D 이미지가 짤린듯한 현사이 발생함

+++ Project_Alice/Project/Alice_Engine/01.Core/Input.cpp
@@ -256,8 +256,8 @@ void CInput::ComputeMouse(bool bStart)
 	GetCursorPos(&tMousePos);
 	ScreenToClient(WINDOWHANDLE, &tMousePos);
 
-	tMousePos.x = tMousePos.x * x;
-	tMousePos.y = tMousePos.y * y;
+	tMousePos.x = (LONG)(tMousePos.x * x);
+	tMousePos.y = (LONG)(tMousePos.y * y);
 
 	if (!bStart)
 	{
